+++ lib/l10n/app_localizations_da.dart
@@ -12974,4 +12974,7 @@ class AppLocalizationsDa extends AppLocalizations {
 
   @override
   String get whatsNewSkipButton => 'Spring over';
+
+  @override
+  String get goalDetailUpdateFailed => 'Seneste opdatering mislykkedes';
 }