+++ lib/l10n/app_localizations_nl.dart
@@ -13014,4 +13014,7 @@ class AppLocalizationsNl extends AppLocalizations {
 
   @override
   String get whatsNewSkipButton => 'Overslaan';
+
+  @override
+  String get goalDetailUpdateFailed => 'Laatste update mislukt';
 }