+++ lib/l10n/app_localizations_sv.dart
@@ -12984,4 +12984,7 @@ class AppLocalizationsSv extends AppLocalizations {
 
   @override
   String get whatsNewSkipButton => 'Hoppa över';
+
+  @override
+  String get goalDetailUpdateFailed => 'Senaste uppdateringen misslyckades';
 }